# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "BalleBaazi 2.0 Apis",
        "description": "# Introduction\nBalleBaazi 2.0 apis on nodejs\n\n# Overview\nCommon header must be passed in every api request\n\n\n# Authentication\nBrrar authentication is applied\n\n# Error Codes\nWhat errors and status codes can a user expect?\nSuccess: 200,\nError: 400,\nMax otp attempts: 401,\nToken_expired: 800,\nNot_found: 404,\nUnauthorized_user: 401\n\n# Rate limit\nIs there a limit to the number of requests an user can send?",
        "contact": {}
    },
    "host": "http://localhost:3000",
    "basePath": "/",
    "securityDefinitions": {},
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/tickets": {
            "get": {
                "summary": "User active tickets",
                "tags": [
                    "users"
                ],
                "operationId": "Useractivetickets",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "default": "Bearer {token}",
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "format": "int32",
                        "description": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                }
            }
        },
...
    },
...
}